revisi total

+++ api/komentar/delete.php
@@ -1,6 +1,8 @@
 <?php
+header("Content-Type: application/json");
+
 include_once '../conf/db_config.php';
-include_once '../../model/Komentar.php';
+include_once '../model/Komentar.php';
 
 $db = new Database();
 $conn = $db->connect();
@@ -8,11 +10,17 @@ $conn = $db->connect();
 $komentar = new Komentar($conn);
 $data = json_decode(file_get_contents("php://input"));
 
-$komentar->id = $data->id;
-$komentar->user_id = $data->user_id;
+if (isset($data->id)) {
+    $komentar->id = $data->id;
 
-if ($komentar->delete()) {
-    echo json_encode(["message" => "Komentar berhasil dihapus"]);
+    if ($komentar->delete()) {
+        http_response_code(200);
+        echo json_encode(["success" => true, "message" => "Komentar berhasil dihapus"]);
+    } else {
+        http_response_code(500);
+        echo json_encode(["success" => false, "message" => "Gagal menghapus komentar"]);
+    }
 } else {
-    echo json_encode(["message" => "Gagal menghapus komentar"]);
+    http_response_code(400);
+    echo json_encode(["success" => false, "message" => "ID komentar tidak dikirim"]);
 }
